updated file structure, updated paths, added scripts, added dependencies, connected front and back end.
