+++ package.json
@@ -4,8 +4,15 @@
 	"description": "",
 	"main": "server.js",
 	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1",
-		"start": "node server.js"
+		"start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev",
+		"start:prod": "node server.js",
+		"start:dev": "concurrently \"nodemon --ignore 'client/*'\" \"npm run client\"",
+		"client": "cd client && npm run start",
+		"seed": "node scripts/seedDB.js",
+		"install": "cd client && npm install",
+		"build": "cd client && npm run build",
+		"heroku-postbuild": "npm run build",
+		"test": "echo \"Error: no test specified\" && exit 1"
 	},
 	"keywords": [],
 	"author": "",
@@ -18,6 +25,7 @@
 		"express": "^4.17.1",
 		"express-ejs-layouts": "^2.5.0",
 		"express-session": "^1.17.1",
+		"if-env": "^1.0.4",
 		"method-override": "^3.0.0",
 		"mongoose": "^5.9.27",
 		"mongoose-permissions": "^1.0.0",
@@ -26,6 +34,8 @@
 		"passport-local-mongoose": "^6.1.0"
 	},
 	"devDependencies": {
-		"jest": "26.6.0"
+		"jest": "26.6.0",
+		"concurrently": "^4.1.0",
+		"nodemon": "^1.18.7"
 	}
 }
